Introduce a basic plugin class to not expose some methods in our API
