+++ spyder/plugins/ipythonconsole.py
@@ -721,7 +721,7 @@ class IPythonConsole(SpyderPluginWidget):
             self.variableexplorer.set_shellwidget_from_id(id(sw))
             self.help.set_shell(sw)
         self.main.last_console_plugin_focus_was_python = False
-        self.update_plugin_title.emit()
+        self.sig_update_plugin_title.emit()
 
     def get_plugin_actions(self):
         """Return a list of actions related to plugin"""
@@ -1117,7 +1117,7 @@ class IPythonConsole(SpyderPluginWidget):
         self.clients.remove(client)
         if not self.tabwidget.count() and self.create_new_client_if_empty:
             self.create_new_client()
-        self.update_plugin_title.emit()
+        self.sig_update_plugin_title.emit()
 
     def get_client_index_from_id(self, client_id):
         """Return client index from id"""
@@ -1317,7 +1317,7 @@ class IPythonConsole(SpyderPluginWidget):
         """
         client = self.clients.pop(index_from)
         self.clients.insert(index_to, client)
-        self.update_plugin_title.emit()
+        self.sig_update_plugin_title.emit()
 
     #------ Public API (for help) ---------------------------------------------
     def go_to_error(self, text):
